create a prompt screen fix ui & add    CircularProgressIndicator

+++ lib/features/promptscreen/bloc/prompt_state.dart
@@ -2,13 +2,13 @@ enum Status { initial, loading, success, error }
 
 class PromptState {
   final Status? status;
-  final int? progress;
+  final double? progress;
   PromptState({this.status = Status.initial, this.progress = 0});
   factory PromptState.initial() {
     return PromptState();
   }
 
-  PromptState copyWith({Status? status, int? progress}) {
+  PromptState copyWith({Status? status, double? progress}) {
     return PromptState(
       status: status ?? this.status,
       progress: progress ?? this.progress,
